fix(appointment): 약속 생성이 되지 않은 문제를 해결

+++ src/main/java/com/swyp3/babpool/domain/possibledatetime/application/PossibleDateTimeServiceImpl.java
@@ -54,17 +54,19 @@ public class PossibleDateTimeServiceImpl implements PossibleDateTimeService{
     public List<PossibleDateTimeResponse> updatePossibleDateTime(Long userId, PossibleDateTimeUpdateRequest possibleDateTimeUpdateRequest) {
         
         // 일정 제거 : possibleDateTimeStatus 가 RESERVED 인 경우는 제외하고 삭제
-        possibleDateTimeRepository.deletePossibleDateTimeWhereStatusIsNotReserved(userId, possibleDateTimeUpdateRequest.getPossibleDateTimeDelList());
-
+        if(!possibleDateTimeUpdateRequest.getPossibleDateTimeDelList().isEmpty()) {
+            possibleDateTimeRepository.deletePossibleDateTimeWhereStatusIsNotReserved(userId, possibleDateTimeUpdateRequest.getPossibleDateTimeDelList());
+        }
         // 추가
-        possibleDateTimeRepository.savePossibleDateTimeList(possibleDateTimeUpdateRequest.getPossibleDateTimeAddList().stream()
-                .map(datetime -> PossibleDateTime.builder().possibleDateTimeId(tsidKeyGenerator.generateTsid())
-                        .userId(userId)
-                        .possibleDateTime(datetime)
-                        .possibleDateTimeStatus(PossibleDateTimeStatusType.AVAILABLE)
-                        .build())
-                .toList());
-
+        if(!possibleDateTimeUpdateRequest.getPossibleDateTimeAddList().isEmpty()) {
+            possibleDateTimeRepository.savePossibleDateTimeList(possibleDateTimeUpdateRequest.getPossibleDateTimeAddList().stream()
+                    .map(datetime -> PossibleDateTime.builder().possibleDateTimeId(tsidKeyGenerator.generateTsid())
+                            .userId(userId)
+                            .possibleDateTime(datetime)
+                            .possibleDateTimeStatus(PossibleDateTimeStatusType.AVAILABLE)
+                            .build())
+                    .toList());
+        }
         // 해당 사용자가 활성화한 모든 가능한 일정 조회해서 응답
         return possibleDateTimeRepository.findAllByUserIdWhereFromThisMonth(userId)
                 .stream()
@@ -88,6 +90,13 @@ public class PossibleDateTimeServiceImpl implements PossibleDateTimeService{
         return possibleDateTimeRepository.findById(possibleDateTimeId).orElseThrow(
                 () -> new PossibleDateTimeException(PossibleDateTimeErrorCode.POSSIBLE_DATETIME_NOT_FOUND, "조회된 PossibleDateTime 이 존재하지 않습니다.")
         ).getPossibleDateTime();
+    }
+
+    @Override
+    public Long findPossibleDateTimeIdByReceiverAndDateTimeAndStatus(Long receiverUserId, LocalDateTime possibleDateTime, PossibleDateTimeStatusType possibleDateTimeStatus) {
+        return possibleDateTimeRepository.findByUserIdAndDateTimeWhereStatus(receiverUserId, possibleDateTime, possibleDateTimeStatus).orElseThrow(
+                () -> new PossibleDateTimeException(PossibleDateTimeErrorCode.POSSIBLE_DATETIME_NOT_FOUND, "조회된 PossibleDateTime 이 존재하지 않습니다.")
+        ).getPossibleDateTimeId();
     }
 
     private void validateRequestPossibleDateTime(Map<String, List<Integer>> possibleDateMap) {
